(김우섭) fix: : "변경사항 임시 저장"
넥스트 페이즈에 작성했던 수정된 로직 페이즈 업데이트 알림으로 이동 - 아직 수정 중

각 카드마다 브레이크 걸기
